chore: rename LoadedNs to Ns
