+++ tests/php/Integration/Run/Command/AbstractTestCommand.php
@@ -19,9 +19,9 @@ abstract class AbstractTestCommand extends TestCase
     {
         $output = $this->createStub(OutputInterface::class);
         $output->method('writeln')
-            ->willReturnCallback(static fn (string $str): void => print $str . PHP_EOL);
+            ->willReturnCallback(static fn (string $str): int => print $str . PHP_EOL);
         $output->method('write')
-            ->willReturnCallback(static fn (string $str): void => print $str);
+            ->willReturnCallback(static fn (string $str): int => print $str);
 
         return $output;
     }
